+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -14,7 +14,7 @@ export class UsersService {
     @InjectRepository(Role) public readonly roleRepository: Repository<User>,
   ) { }
   
-  async create(createUserDto: CreateUserDto) {
+  async create(createUserDto: CreateUserDto): Promise<User> {
     const roleReq = await this.roleRepository.findOne(createUserDto.roleId,  {
       relations: ['users']
     })
@@ -24,27 +24,37 @@ export class UsersService {
     console.log( "User DTO : "+ (createUserDto instanceof CreateUserDto) + " = " + JSON.stringify(createUserDto));
     console.log( "User object : "+ (user instanceof User) + " = " + JSON.stringify(user));
     console.log( "requested role : "+ (roleReq instanceof Role) + " = " + JSON.stringify(roleReq));
-    return 'This action adds a new user' + JSON.stringify(createUserDto);
+    return user;
   }
 
-  async findAll() {
+  async findAll(): Promise <User[]> {
     const users = await this.repository.find();
     console.log("Loaded users: " + JSON.stringify(users) )
     return users;
   }
 
-  async findOne(id: number) {
+  async findOne(id: number): Promise <User> {
     const user =  await this.repository.findOne(id, {
       relations: ['role']
     });
+
+    if (!user) {
+      throw new NotFoundException(`User ${id} not found`);
+    }
+
     console.log("User " + id + " is: " + JSON.stringify(user) )
     return user;
   }
 
-  async update(id: number, updateUserDto: UpdateUserDto) {
+  async update(id: number, updateUserDto: UpdateUserDto): Promise <User> {
     const user =  await this.repository.findOne(id, {
       relations: ['role']
     }); 
+
+    if (!user) {
+      throw new NotFoundException(`User ${id} not found`);
+    }
+
     console.log("DTO  " + JSON.stringify(updateUserDto));
     user.firstName = (updateUserDto.firstName == undefined) ? user.firstName : updateUserDto.firstName
     console.log(`firstName ${updateUserDto.firstName}`);
@@ -53,12 +63,19 @@ export class UsersService {
     user.eMail = (updateUserDto.eMail == undefined) ? user.eMail : updateUserDto.eMail
     console.log(`eMail ${updateUserDto.eMail}`);
     console.log(`roleId ${updateUserDto.roleId}`);
-    await this.repository.save(user);
-    return `This action updates a #${id} user`;
+    return this.repository.save(user);
   }
 
-  async remove(id: number) {
+  async remove(id: number): Promise<User> {
+    const deletedUser =  await this.repository.findOne(id, {
+      relations: ['role']
+    });
+
+    if (!deletedUser) {
+      throw new NotFoundException(`User ${id} not found`);
+    }
+    
     await this.repository.delete(id);
-    return `This action removes a #${id} user`;
+    return deletedUser;
   }
 }
